Experimental cmake files.
cmake (www.cmake.org) is a build system - it's one of the alternatives currently
being investigated as autotools replacement at the KDE project (like
scons/bksys).

cmake is actually a Makefile generator. The ./configure ; make equivalent is:
ccmake
-> press "c"
-> press "c"
-> press "g"
make

ccmake is a console (not command line) interface to configure the program.
Probably there is a command line equivalent to the above, but I haven't found it
yet.

The cmake port was pretty easy: it got completed within a single day (I ported 2
of my projects within that time).
A big advantage to scons/bksys is, that cmake is much more mature: many issues
of scons/bksys (such as the configuration architecture) is solved in cmake
already or at least in a much better state. Also it does not suffer from the
same performance problems as scons: with scons _every_ run of scons (the "make"
equivalent) takes several seconds or even minutes before it does anything
useful. cmake only generates the Makefiles and therefore does not have this
problem by design. Note that the problem of scons/bksys is being worked on and
with "waf" apparently solved already.

However there are also disadvantages with cmake:
* The inherent problems of the make and friends are not solved.
  It still isn't completely clear to me when to use "ccmake" again and
  reconfigure the project / regenerate the Makefiles and when a simple "make" is
  sufficient.
* Using the same source files in two different targets will cause cmake to
  compile the sources twice.
  This can be solved using a static library in between - but this isn't pretty.
* cmake must be downloaded, compiled and (optionally) installed before compiling
  boson can even start.
  This problem does not exist with scons/bksys at all, since we can easily
  provide a copy of it (it's very lightwight). The additional python requirement
  of scons is not a problem, as we need that anyway.


It seems that boson can live with both, scons/bksys/waf and cmake. We will
probably choose whichever KDE will choose. For now scons remains my primary
build system, as it completely works currently.
The "make install" of cmake has not yet been worked on.