allow to change the "side" of the players. in the newgame widget the players now have a "Side" combobox that they can select from, containing "Player 1" ... ("Player %1").arg(maxPlayers) and "Watch only".
The "Watch only" adds a player that won't be able to control any units and won't
be relevant for the game in any way - but it will be able to watch the game.
This is an experimental feature only though. Also the player currently won't see
anything onless he does a debug->unfog :-)