Removing all autotools files.
I had planned to support autotools for at least one more release, however it
appears to be more work than it is worth - so we simply drop it now. We wanted
to do so anyway.

Currently the only way to build boson from svn trunk is scons/bksys, see
boson.eu.org (thanks, Rivo!)

this commit also removes the "admin" external and the svn:ignore properties for
autotools.