install the species.
the actual installation code is completely in the BOSON_INSTALL_SPECIES macro,
not in the actual CMakeLists.txt files.
this way we need to write the code only once and can use it for all species.

some things are not "nice" here - I'd prefer to install the complete directories
sometimes (e.g. units) but cmake doesn't provide a way to do so. instead we now
install all relevant files from these directories (in the case of units:
index.unit, unit.3ds *.png).

in addition to the actual installation code, the BOSON_INSTALL_SPECIES macro
also checks whether the IDs of the units are valid, i.e. that a certain ID only
appears once and that every index.unit file actually has an ID.
this woks only when cmake is run (which is not necessary when adding a unit!)
however I think we can live with this limitation :-)