provide a copy of lib3ds
this means for now we don't depend on lib3ds anymore.
this has become necessary, because gcc 4.x miscompiles lib3ds 1.2.0 which still
is the most recent released version. the bug is fixed in lib3ds cvs and several
distributions, but i'm sure many people will fall into this trap when we release
0.12.

we might remove this copy again once a new (fixed) lib3ds version is released,
so that we can depend on that version.