set the CMAKE_BUILD_WITH_INSTALL_RPATH variable. this avoids relinking at installation.
we link statically against our libs, so there is nothing that rpath helps
finding.

there _might_ be one exception: with rpath being disabled it will be necessary
to install plugins before starting boson. this has always been the case before,
so it isn't a hard restriction.
also i am not sure whether the rpath actually covers plugin loading: it might be
that our prefix is more or less hardcoded into the plugin loading code. if
anyone is interested in this, you can try whether boson can find its plugins
when they are not installed, if rpath is enabled. if it does, we should probably
re-enable the rpath thing.