make sure that the size of the array in makeCells() never becomes negative (or very large).
this was possible before, when a unit went off the map, which currently happens
a lot when destroying a flying unit (they don't stop moving when being
destroyed).

this fixes a "happens sometimes" crash which i have experienced a lot in the
past, but was (until now) never able to make reproducable and fix it.