adding config scripts: a config script is a collection of config entries (keys) with a value assigned. whenever the script is executed, the current config values are replaced by that value. this in particular can be used to provide "profiles" for e.g. rendering performance.
the OptionsDialog makes use of this for the renderingspeed setting now, but
won't actually execute the scripts: instead it'll "peek" at the values in the
script and set the widgets accordingly.

this was primarily meant to make "profiles" much easier: instead of having to
edit the optionsdialog, hoping to catch all the values correclty, we simply edit
the script now. the rest should work mostly automatically. however if a config
widget uses a custom widget (not a ConfigOptionWidget), then still some work is
required (but less than before) - the value of a script must be assigned to
that widget.

as a little bonus, we could even execute the scripts on startup if desired, e.g.
provide a --fastrendering command line switch, which would behave pretty much
exactly like the optionsdialog