adding a file converter program. this program simply takes a .bpf/.bsg file and converts it to the most current file format.
it does exactly what boson is doing when loading such a file.

when converting a file from an older format, the preferred way to do so is to
use this program. it simply reads the tar archive, converts the files in it, and
writes the tar archive back to a (new) file.
when loading a file using the map editor and immediately saving it again, boson
also loads the files from the tar archive into it's own data structures which is
clearly much more error prone: when bofileconverter converts something wrong,
boson will do so, too. but when boson does something wrong, the fileconverter
may still work correctly (and thus the "converted" file won't be broken
afterwards)