Adding an improved cmake script to find the python libraries: use the python executable to find out which libraries the python libraries depend on and add these libs to PYTHON_LIBRARIES.
This should fix the problems where linking boson failed because of undefined
references caused by python.
Note that this error appears on "some" systems only, as the dependency on other
systems (e.g. all that we tested on prior to releasing 0.12) have kdelibs or
some other library depend on these libraries, too.

In addition this script honors the REQUIRED and the QUIET flags of
find_package().


note that (in contrast to cmake's original FindPythonLibs.cmake) this script has
been tested on and probably works under unix only.