Map previews.
Every map file (.bpf/.bsg) can now have a mappreview/map.png file, containing a
small preview of the map (usually the ground in the minimap). This file is
supposed NOT to contain any units: we can add a
mappreview/units_that_player_n_can_see.png file for that.

This file is meant to be displayed in the newgame widget, if that file (which is
optional) is present. If it is not available, there is simply no preview being
displayed (calculating one on startup would be way too time consuming).