adding a small program to add mappreview/map.png files to the .bpf/.bsg files.
this program takes two .bpf files (input and output) and behaves pretty much
exactly like bofileconverter: it reads/untars the input file (and of
course automatically converts it to most current file format) and writes it back
to the output file.
but in addition it also loads the data into a BosonPlayField object and
retrieves a mappreview from there that is written, too.

the BosonPlayField is never saved, it is used to retrieve the map preview only,
so there won't be any additional bug source here. actually bofileconverter could
be replaced completely by bocreatepreview easily.