Minimap redesign.
Both, actual ideas and concrete implementations have been preserved nearly
completely - only the class design has been changed. The "renderer" approach has
been removed and has been replaced by proper use of a BoUfo class.
Also the UDrawable is not used anymore - instead we use the (newer, simpler,
more powerful) BoUfoCustomRenderer.

We now have a 3 level class design:
1. BosonGLCompleteMiniMap
This is a minimap of the complete map. All textures at full size are maintained
here. Rendering happens to a quad of size (1, 1).

2. BosonGLMiniMapView
A view of BosonGLCompleteMiniMap. This is meant to display only a part of the
complete minimap, probably at some fixed (screen-)size. (Mostly
not-yet-implemented though)
I also decided to render the camera in this class - the camera is a border case
and would (from a design pov) be acceptable in both, the View and the Complete
class.

3. BosonUfoMiniMap
This is the actual minimap interface to the outside world. It is a BoUfo widget
around the minimap that provides additional features, such as
* a logo when the player has no minimap
* zoom in/out buttons (the images are implemented, the actual button semantic
  not yet)
* mouse event handling


Most "interesting" code should usually go to BosonGLCompleteMiniMap, whereas
non-minimap classes should use BosonUfoMiniMap exclusively.