moving the minimap dir to gameview
originally i planned to use the minimap in the startup widgets, however the
intention of the minimap in the game differs totally from the intention in the
startup widgets, and thus the API should be completely different, too. therefore
it made much more sense to create a dedicated mappreview widget for the
startup widgets instead.