adding --fast and --veryfast command line options
these are meant to make startup faster (not the game itself):
--fast should make it as fast as possible without sacrificing important parts
of the game. atm this means it is equal to --nosound and --notexturecompression.
--veryfast should make it as fast as possible ignoring any losses. this is meant
as debugging mode. atm it is equal to --fast and --nomodels.

this should make my command lines a lot shorter: instead of using
  ./boson --new --nosound --nomodels --notexturecompression --start
i can use
  ./boson --new --veryfast --start
which clearly is a lot shorter