adding a BoGL class
this class does pretty much everything that the bogl* functions have done so far
(they are still provided, too). but a class makes it much easier to store some
data and add new functionality.

in particluar the file names that are used for libGL.so/libGLU.so are now stored
and can be retrieved at any time.