adding a workaround for mesa <= 6.5.1 - disable vertex arrays.
in mesa <= 6.5.1 glPopClientAttrib() has a pretty dramatic bug which causes it
to crash - that function is used mostly when using vertex arrays.
when we disable vertex arrays, we can avoid this crash.

in particular this means completely disabling water rendering, as we dont have a
non vertex array version of it.