upheap(x) and downheap(x) with x>= heap_size must be noops.
this is actually (implcitly) assumed by (at least) MxHeap::remove(), as it calls
upheap(x) where x being the old (!) position of the removed item. this is
allright as long as x was not == length()-1, i.e. not the last item in the heap.
if it was the last item however, we'll call upheap/downheap(x) with
  x==length()
i.e. oldlength() - 1 (as length() has been reduced by 1 because the item was
removed).

if the upheap()/downheap() is now actually executed, then the now out-of-bounds
item is (usually) moved down to a valid index again and therefore _not_ removed
from the heap.

this causes memory corruption and usually a segfault afterwards, if the item
that should have been removed gets deleted afterwards.

this happens on at least one AMD64 machine. this problem might(!) actually be
related to a different problem, as apparently the bobmfconverter behaves
differently on this machine (it never tries to remove the last item of a heap on
my machine).
so it may be possible that the generated model file (more exactly: the LODs)
will be broken/look weird. but maybe we're lucky - we'll see.

however the program should not crash anymore now.

thanks a lot to goneri who helped greatly at tracking this down