splitting up unitplugins.* into several files
the file unitplugins/unitplugins.h can be used as a replacement for the previous
unitplugins.h: it simply includes the headers of all plugins and thus keeps the
visible API intact.
however it may be better to include the required plugins only in order to keep
recompilations to a minimum.