Adding "units inside units" support. WARNING: I am going to write a pretty long log here, because I believe the "history" aspects of this feature actually belong to the "svn history", not to the wiki/mailinglist/whatever.
Short introduction:
"UnitsInsideUnits" is a feature that is meant to implement
* transportational units
* units "walking out" of factories and the like
* airports/helipads
* any other kind of facility that might need to store a unit temporarily
  (ammunition storage, repaid pad, ...)

in theory this feature could even be used to implement things like bridges
(simply create a unit called "bridge" and force any unit entering it to
immediately move out of it again on the other side).

Obviously supporting all these features is a rather big job.
The current implementation supports most of these, with some restrictions
* storages must not be mobile (i.e. no transportational units)
* storages should not be rotated (see below for details)
* at most one unit at a time may enter/leave the storage


There are two parts in this patch:
1. gameengine
2. user interface (gameview)

The second part obviously depends on the first and is straight forward, I don't
have anything to say on this.

The central classes of gameengine part are
* EnterUnitPlugin
* UnitStoragePlugin
Every mobile unit now has a EnterUnitPlugin and thus can in theory enter a unit
storage. The plugin handles most of the actual entering.
A unit with a UnitStoragePlugin is a "unit storage" - units can enter it. This
plugin is responsible for the decision which unit is allowed to enter it and
which is not (it can approve and deny "enter permissions").

Both plugins are essentially state machines (this is the reason for the "at most
one unit at a time may enter/leave" restriction - allowing multiple units would
make the state machine much more complicated). The concrete details of these
state machines can atm be found at
http://boson.freehackers.org/wiki/Main/UnitsInUnits

The EnterUnitPlugin is also responsible for leaving a unit, i.e. not only for
the task "entering" but for everything that is related to a unit that enters
another unit.

The UnitStoragePlugin primarily contains the "paths" - they define how, where,
which unit may enter the storage. At most one unit at a time may be on a path.
The EnterUnitPlugin requests such a path from the storage and (if it receives
it) uses it as a replacement for the pathfinder in its UnitMover.

Note that the current implementation is completely experimental. In particular
the fact that entering/leaving can be aborted at any time, but the plugins won't
get notified, is an extremely critical open issue: you can easily render both
units useless because of it (enter a storage with a unit ; move it out again
before it arrived its final destination - it is now stuck and both units are
useless).




So much for the overview - the rest of this commit log is mostly details and
some history.

more detailed NOT yet supported list:
- unit storage plugin for a mobile unit - i.e. transportation units. main
  remaining problems here:
  - units inside the storage must move along with the unit
  - units inside the storage probably must be invisible (easy to solve)

  I expect this not to be too difficult to implement, however it is not of any
  priority to me, therefore I won't do that fo rnow.
- probably problematic: a unit storage that is rotated.
  -> this issue reveals a few major flaws in current boson design: we pretty
     much ignore the rotation of units completely, in particular at collision
     detection. however that is not acceptable with unit storages, as the unit
     must enter at the side of the storage where the door is.
     right now the code simply assumes that the storage is NOT rotated.
     if it is, weird things can happen (e.g. unit cannot enter at all, due to
     collision detection issues, and unit enters at the wrong side of the
     storage)

  I think the design flaws should be fixed, before this problem should be
  addressed. therefore e.g. removing Unit::x()/y() and replacing it by
  centerX()/centerY() everywhere ; using both, centerX/Y() and rotation values
  in the bounding rectangle ; and things like that.
- aircrafts can "land", but they do so rapidly: while landing, its z
  position remain the same all the time, once it has landed, its z position is
  set to the ground immediately.
  a simple linear interpolation should be pretty easy to implement.

  I might do that "some day soon", but right now I am more concerned with the
  principle functionality, especially aborting enter/leave orders. if anyone
  wants to fix this z issue, feel free to do so.


Some words about the history of the current implementation:
Actually this is already a rewrite. I started out by implementing a much smaller 
subset of the above features (in particular: simply entering and possibly
leaving a facility using a ground unit). The plan was to extend the initial
implementation to (at least) aircrafts entering an airport.
However it soon turned out that such a solution could hardly be extended. Even
the "simple" difference of "airplane" vs "ground unit" entering a storage makes
a huge difference already. At some point the code became so complex that I
decided for a rewrite of most algorithms/code.
-> Really only the algorithms/code. NOT the class design. Actually the current
   class design hasn't changed at all since the very first version, I believe it
   has proven to be very useful and versatile. Only the UnitStoragePlugin
   received a more sophisticated "Path" class at some point.

It also turned out very soon, that the rewrite probably won't succeed "good
enough" for me either: there were simply _way_ too many special cases to think
about (even when restricting to ground and air units entering a facility only!)
Therefore I started to draw some pictures first: I already had a list of
features that entering/leaving could be used for (transportation ; factories ;
bridges ; ...) and now I did a Gantt-chart like drawing for the airplane case.
The result looked a lot like a state machine for both plugins would be the most
appropriate choice.
After I drew a few versions of these (only the "enter" task), I actually came up
with one that I believed could be used for most if not all features described
above. The "leave" state chart was pretty much straight forward once the "enter"
problem was solved.

This actually turned out to be a very good approach: the implementation of the
state machine according to the diagrams was rather simple. Once it was done, I
did the same for the UnitStoragePlugin.
Since the EnterUnitPlugin was already mostly implemented, that was pretty
simple, as I already had designed large parts of the API (in order to use it in
the EnterUnitPlugin).