WARNING: hack!
making MoveInsideUnit uninterruptible. both, replaceToplevelOrder() and
clearOrders() will _not_ remove the current order if it (or its parent) is a
MoveInsideUnit order - these must not be interrupted (otherwise the
entering/leaving  and the storing unit will both be unusable).

this is a very ugly check which is required at all places that might remove the
current order (except currentSuborderDone(), which is allowed to do so). atm
only the above 2 methods do that (+ currentSuborderDone()).
a nicer solution would be some kind of "interruptible" flag in the UnitOrder
class - however this flag must be derived by suborders: if MoveInsideUnit
triggers a Turn suborder, then the Turn suborder must not be aborted either.


this fixes the last major problem that i am aware of with the UnitsInUnits
patch - when a entering/leaving unit is given another order, it won't be in an
unusable state anymore now.
all remaining things are imho "would be nice to have" todos only.