Adding a "bogl compiler" (boglc), which takes the *.boglc files from the specs/ directory and generates *_decl_p.h files from them.
I haven't been able to figure out a way to auto-generate them using cmake (we
need to depend on bogl.h, which causes problems), so new files must be
semi-manually created by compiling boglc and calling make_headers.sh. the
generated files are meant to go to svn then.

generating the headers from *.boglc files has a major advantage: it is _much_
easier to change the format of these files: every function (a few hundred in
total) is used in at least 3 places (typedef, function pointer, #define), so
changings something in the files is a lot of work usually.

now every function appears exactly once in a .boglc file and only the compiler
makes several lines in the header out of it.