- removing preLoadPlayField(), isPreLoaded(), mapPreviewPNGData() and other   methods that have been obsoleted by BPFPreview from BosonPlayField - removing the BPFPreview object from BosonPlayField. the preloaded objects are   stored in BosonData globally, there is no need for another object in   BosonPlayField anymore.   as a consequence: re-introducing the BPFDescription object into   BosonPlayField, the only preview feature that really is required in   BosonPlayField (for the map editor at the very least)
the BosonPlayField class now stores the map and the description, nothing else.
whether it should also store rules and winning conditions one day is still an
open question (but I think it should not)

this also makes the playfieldtest pass, as saving a playfield now provides
exactly the data needed for loading a playfield and no less.