removing BosonItem::x() and BosonItem::y(). replacing all occurances by leftEdge()/topEdge() or centerX()/centerY() depending on the context.
we may re-introduce x() and y() at some point with different semantics (they
should remove the _center_ not the top/left edge).

furthermore, some methods have been renamed - some kind of "TopLeft" or "Center"
suffix has been aded. these functions include:
- canPlaceUnit()
- move()
- createNewItem()
- createItem()
- buildProducedUnit()
- editorPlaceUnit()
- newShot()
(from various classes)

this should make it cleared what kind of parameter the function expects.

the eventual goal is to use the _center_ point to describe the position of an
item whenever possible. this makes much more sense, is much more intuitive and
removes a lot of complications (a lot of code had to convert x() to the center
of the item anyway).

in particular this is required to make collision detection honor the rotations
of items.