test construction of a facility.
when the construction is completed and event is sent through
boGame->eventManager(), therefore we _need_ a Boson object here. for this, a
BosonContainer has been added which simplifies the management of such an object
and should behave essentially like a CanvasContainer.

an event loop or a QApplication still has not been added, so no playerinput is
possible.