port all plugins.
the plugin architecture has caused some significant headache, therefore plugins
are disabled for now, the code is compiled directly into the binary.
this is likely to be the final solution, as the plugins have been meant as a
debugging shortcut only, they never provided (and never will) any additional
features.

boson compiles, links and starts now. you won't see anything, but it starts ;-)